Refactor key into helix-view
Now also make use of Deserialize for Config.
