+++ helix-term/src/main.rs
@@ -1,9 +1,8 @@
+use anyhow::{Context, Error, Result};
 use helix_term::application::Application;
 use helix_term::args::Args;
 use helix_term::config::Config;
 use std::path::PathBuf;
-
-use anyhow::{Context, Result};
 
 fn setup_logging(logpath: PathBuf, verbosity: u64) -> Result<()> {
     let mut base_config = fern::Dispatch::new();
@@ -89,12 +88,11 @@ FLAGS:
         std::fs::create_dir_all(&conf_dir).ok();
     }
 
-    let config = std::fs::read_to_string(conf_dir.join("config.toml"))
-        .ok()
-        .map(|s| toml::from_str(&s))
-        .transpose()?
-        .or_else(|| Some(Config::default()))
-        .unwrap();
+    let config = match std::fs::read_to_string(conf_dir.join("config.toml")) {
+        Ok(config) => toml::from_str(&config)?,
+        Err(err) if err.kind() == std::io::ErrorKind::NotFound => Config::default(),
+        Err(err) => return Err(Error::new(err)),
+    };
 
     setup_logging(logpath, args.verbosity).context("failed to initialize logging")?;
